feat(footer): add footer
Footer has been added and made responsive.

Green color added to theme.

Refactor AppProviders to only contain global providers and not anything detailed.

Grid area has been added for root div. This is in order to structure the page and always keep the navbar at the top and footer at the bottom.