feat(search): add min & max calories inputs & functionality
Everytime you visit the search page, sessionStorage gets cleared.

In Recipes page, client util is being used instead of manually fetching.

In order to display the calories properly, we need to divide it by the yield, because the calories are counted in total, which means the calories for every serving in total.